Substantial Changes
Added an alternate version of the 'cab_tokenizer' function ('cab_tokenizer_no_stop') because it is called directly in 'calculate_tf_idf_terms' function. When called from this function it circumvented the conditional parameter in the 'preprocess_movie_data' function, resulting in all stop words always being stripped which is not the intended effect at 3.2.

'cab_tokenizer' function adjusted to contain the stopwords in the definition. This was in an attempt to trouble-shoot the term 'film' being included in the analysis despite being in the stopwords list.
Later, I realised this is because the stopwords are filtered out BEFORE the tokens are lemmatised, meaning that alternate forms of the word (e.g. 'films', 'filming') would be would not be filtered out and their lemma ('film') would then be included as a toke. 

The 'English' stopwords are stored as a separate variable to the identified corpus-specific stopwords, each token is checked against each set (and the punctuation set) to generate a boolean variables; if any of the variables are true, the token is neither lemmatised nor vectorised.


In the 'preprocess_movie_data' function, the 'default' processing now includes a print statement to alert the user that stopwords are not being filtered out and, implicitly, terms are not being filtered by document frequency.